+++ src/test/java/org/wikiclean/WikiCleanDeTest.java
@@ -27,13 +27,14 @@ import junit.framework.JUnit4TestAdapter;
 import org.apache.commons.io.FileUtils;
 import org.junit.Test;
 import org.wikiclean.WikiClean.WikiLanguage;
+import org.wikiclean.languages.German;
 
 public class WikiCleanDeTest {
 
   @Test
   public void testId1() throws Exception {
     String raw = FileUtils.readFileToString(new File("src/test/resources/dewiki-20130602-id1.xml"), "UTF-8");
-    WikiClean cleaner = new WikiClean.Builder().withLanguage(WikiLanguage.DE).build();
+    WikiClean cleaner = new WikiClean.Builder().withLanguage(new German()).build();
     String content = cleaner.clean(raw);
     //System.out.println(content);
 
@@ -51,7 +52,7 @@ public class WikiCleanDeTest {
   @Test
   public void testId5() throws Exception {
     String raw = FileUtils.readFileToString(new File("src/test/resources/dewiki-20130602-id5.xml"), "UTF-8");
-    WikiClean cleaner = new WikiClean.Builder().withLanguage(WikiLanguage.DE).build();
+    WikiClean cleaner = new WikiClean.Builder().withLanguage(new German()).build();
     String content = cleaner.clean(raw);
     //System.out.println(content);
 
@@ -68,7 +69,7 @@ public class WikiCleanDeTest {
   @Test
   public void testId81() throws Exception {
     String raw = FileUtils.readFileToString(new File("src/test/resources/dewiki-20130602-id81.xml"), "UTF-8");
-    WikiClean cleaner = new WikiClean.Builder().withLanguage(WikiLanguage.DE).build();
+    WikiClean cleaner = new WikiClean.Builder().withLanguage(new German()).build();
     String content = cleaner.clean(raw);
     //System.out.println(content);
 
@@ -81,7 +82,7 @@ public class WikiCleanDeTest {
   @Test
   public void testId89() throws Exception {
     String raw = FileUtils.readFileToString(new File("src/test/resources/dewiki-20130602-id89.xml"), "UTF-8");
-    WikiClean cleaner = new WikiClean.Builder().withLanguage(WikiLanguage.DE).build();
+    WikiClean cleaner = new WikiClean.Builder().withLanguage(new German()).build();
     String content = cleaner.clean(raw);
     //System.out.println(content);
 
@@ -92,7 +93,7 @@ public class WikiCleanDeTest {
   @Test
   public void testId111() throws Exception {
     String raw = FileUtils.readFileToString(new File("src/test/resources/dewiki-20130602-id111.xml"), "UTF-8");
-    WikiClean cleaner = new WikiClean.Builder().withLanguage(WikiLanguage.DE).build();
+    WikiClean cleaner = new WikiClean.Builder().withLanguage(new German()).build();
     String content = cleaner.clean(raw);
     //System.out.println(content);